% 
% Copyright (c) 2010, 2014, IBM Corp. All rights reserved. 
%   	
% This program is free software: you can redistribute it and/or modify 
% it under the terms of the GNU General Public License as published by 
% the Free Software Foundation, either version 3 of the License, or 
% (at your option) any later version. 
%
% This program is distributed in the hope that it will be useful, 
% but WITHOUT ANY WARRANTY; without even the implied warranty of 
% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the 
% GNU General Public License for more details. 
%
% You should have received a copy of the GNU General Public License 
% along with this program. If not, see <http://www.gnu.org/licenses/>. 
% 
% 
\name{nzArule} 
\Rdversion{1.0}
\alias{nzArule}
 
\title{Wrapper for Association Rules (ARules)}
\description{
	This function is a wrapper for the \code{nza..ARULE} stored procedure.
	All computations are performed on the NPS, the results are converted to a R object.

	A local version of an association rule algorithm may be found in \code{arules::apriori}.
}
\usage{
nzArule(
		data,
		tid  = "tid",
		item = "item",
		by = NULL,
		lvl = 1,
		maxsetsize = 6,
		support = 5,
		supporttype = "percent",
		confidence = 0.5,
		model = NULL,
		format = "arule",
		dropAfter = F,
		...
)
}
\arguments{
	\item{data}{A \code{nz.data.frame} object storing the data to be mined.}
	\item{tid}{Input table column that identifies the transactions by an id.}
	\item{item}{Input table column that identifies items in transactions by an \code{id} or \code{string}.}
	\item{model}{Optional: The modelname. Will be autogenerated if not specified.}
	\item{format}{May be \code{arule} for output compatible to packages 'arules' and 'arulesViz' or \code{raw} for downloading the raw results from the database.}
	\item{dropAfter}{\code{TRUE} to drop all created tables after all results have been fetched.}
	\item{by, lvl, maxsetsize, support, supporttype, confidence, ...}{Further arguments to be passed to \code{nza..ARULE}.}
}
\description{
	There are notable overloaded functions such as \code{print} and \code{plot} for objects of class
	\code{rules} available e.g. in \code{arules} or \code{arulesViz}.
	
	For more details on the \emph{ARULE} algorithm and requirements to the \code{data}, please refer to
	the documentation of \code{nza..ARULE} in the Netezza
	In-Database Analytics Reference Guide or Netezza In-Database Analytics Developers Guide.
}
\value{Object of class \code{rules} or list of \code{data.frame}s}
\note{Please note that for arules plotting, you will need to install and load the 'arulesViz' package.}

\examples{\dontrun{
	require(nza)
	
	# NZSQL - ODBC connection id, has to be defined locally 
	# on your system using "ODBC_Source_Administrator" 
	# nzConnectDSN('NZSQL') 
	
	# Sample taken from arules' GROCERIES data
	groceries <- nz.data.frame("GROCERIES") 
	nzArule(groceries, "TID", "ITEM")
	
	# nzDisconnect() 
}
}
